moves "all positions" component, fixes imports
