+++ frontend/src/Pages/OrganizationRecruitmentPage/Components/GangPositionDropdown/GangPositionDropdown.module.scss
@@ -1,6 +1,6 @@
-@import 'src/constants';
+@import '../../../../constants';
 
-@import 'src/mixins';
+@import '../../../../mixins';
 
 .container {
   display: flex;
